+++ src/main/java/com/mraof/minestuck/inventory/captchalogue/SetModus.java
@@ -1,6 +1,6 @@
 package com.mraof.minestuck.inventory.captchalogue;
 
-import com.mraof.minestuck.MinestuckConfig;
+import com.mraof.minestuck.entity.MSAttributes;
 import com.mraof.minestuck.item.CaptchaCardItem;
 import com.mraof.minestuck.item.MSItems;
 import com.mraof.minestuck.util.MSSoundEvents;
@@ -8,6 +8,7 @@ import net.minecraft.core.HolderLookup;
 import net.minecraft.core.NonNullList;
 import net.minecraft.nbt.CompoundTag;
 import net.minecraft.nbt.Tag;
+import net.minecraft.network.chat.Component;
 import net.minecraft.server.level.ServerPlayer;
 import net.minecraft.sounds.SoundSource;
 import net.minecraft.world.item.ItemStack;
@@ -56,8 +57,8 @@ public class SetModus extends Modus
 		list = NonNullList.create();
 		
 		for(int i = 0; i < size; i++)
-			if(nbt.contains("item"+i, Tag.TAG_COMPOUND))
-				list.add(ItemStack.parse(provider, nbt.getCompound("item"+i)).orElseThrow());
+			if(nbt.contains("item" + i, Tag.TAG_COMPOUND))
+				list.add(ItemStack.parse(provider, nbt.getCompound("item" + i)).orElseThrow());
 			else break;
 		if(side == LogicalSide.CLIENT)
 		{
@@ -75,7 +76,7 @@ public class SetModus extends Modus
 		for(int i = 0; i < list.size(); i++)
 		{
 			ItemStack stack = iter.next();
-			nbt.put("item"+i, stack.save(provider));
+			nbt.put("item" + i, stack.save(provider));
 		}
 		return nbt;
 	}
@@ -110,7 +111,7 @@ public class SetModus extends Modus
 	@Override
 	public NonNullList<ItemStack> getItems()
 	{
-		if(side == LogicalSide.SERVER)	//Used only when replacing the modus
+		if(side == LogicalSide.SERVER)    //Used only when replacing the modus
 		{
 			NonNullList<ItemStack> items = NonNullList.create();
 			fillList(items);
@@ -137,8 +138,11 @@ public class SetModus extends Modus
 	@Override
 	public boolean increaseSize(ServerPlayer player)
 	{
-		if(MinestuckConfig.SERVER.modusMaxSize.get() > 0 && size >= MinestuckConfig.SERVER.modusMaxSize.get())
+		if(hasHitMaxCards(player, size))
+		{
+			player.displayClientMessage(Component.translatable(CAPTCHA_LIMIT), true);
 			return false;
+		}
 		
 		size++;
 		markDirty();